requirements.txt
fastapi
uvicorn
imagededup
python-multipart
Pillow
pillow-heif
gunicorn